[AIE][TableGen] Factor out ConstTable.h
